🏗 create builtins architecture
