+++ minishell/Sources/builtins/cd.c
@@ -1,26 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   3_builtins.c                                       :+:      :+:    :+:   */
-/*   By: sbouras <sbouras@student.42quebec.com>       +:+ +:+         +:+     */
-/*   By: mdoquocb <mdoquocb@student.42quebec.com>   +#+  +:+       +#+        */
+/*   cd.c                                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tonted <tonted@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/11 18:29:46 by mdoquocb          #+#    #+#             */
-/*   Updated: 2022/06/13 15:10:10 by jrossign         ###   ########.ca       */
+/*   Created: 2022/09/11 09:08:33 by tonted            #+#    #+#             */
+/*   Updated: 2022/09/11 09:08:53 by tonted           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minishell.h>
-
-static int	exec_pwd(void)
-{
-	char	*buf;
-
-	buf = getcwd(NULL, 0);
-	printf("%s\n", buf);
-	free(buf);
-	return (1);
-}
+#include "minishell.h"
 
 static int	exec_cd2(char **str, char ***envp)
 {
@@ -48,7 +38,7 @@ static int	exec_cd2(char **str, char ***envp)
 	return (1);
 }
 
-static int	exec_cd(char *dir, char ***envp)
+int	exec_cd(char *dir, char ***envp)
 {
 	char	*str;
 	int		i;
@@ -65,31 +55,4 @@ static int	exec_cd(char *dir, char ***envp)
 	if (chdir(dir))
 		return (perror_minishell(errno, "minishell:"));
 	return (exec_cd2(&str, envp));
-}
-
-static int	exec_env(char **envp)
-{
-	while(*envp)
-	{
-		printf("%s\n", *envp);
-		envp++;
-	}
-	return (1);
-}
-
-int	exec_builtins1(t_cmd *cmd, int *status, char ***envp, int process)
-{
-	int	i;
-
-	i = 0;
-	if (!ft_strncmp(cmd->cmd[0], "cd", 2))
-		return (exec_cd(cmd->cmd[1], envp));
-	if (!ft_strncmp(cmd->cmd[0], "pwd", 3))
-		return (exec_pwd());
-	if (!ft_strncmp(cmd->cmd[0], "env", 3))
-		return (exec_env(*envp));
-	else
-		i = exec_builtins2(cmd, status, envp, process);
-	return (i);
-	
 }
