+++ minishell/Sources/builtins/env.c
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tonted <tonted@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/11 09:09:21 by tonted            #+#    #+#             */
+/*   Updated: 2022/09/11 09:09:30 by tonted           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	exec_env(char **envp)
+{
+	while(*envp)
+	{
+		printf("%s\n", *envp);
+		envp++;
+	}
+	return (1);
+}
