+++ minishell/Sources/builtins/pwd.c
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tonted <tonted@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/11 09:07:38 by tonted            #+#    #+#             */
+/*   Updated: 2022/09/11 09:08:17 by tonted           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	exec_pwd(void)
+{
+	char	*buf;
+
+	buf = getcwd(NULL, 0);
+	printf("%s\n", buf);
+	free(buf);
+	return (1);
+}
